REFACTOR AuthenticatedLayout and UnauthenticatedLayout to separate files
